* master: (34 commits)
  update changelog
  fix for column alignment for standalone accessors
  Disable the REST example in build.sh
  renamed defaultPath to defaultRealmPath and removed unused/unexposed setDefaultRealmPath:
  Added [RLMRealm defaultRealmPath]
  Remove CocoaPod buttons
  clarify core licensing
  Fixed how Xcode6 iOS framework is combined into a fat framework
  Use of tightdb::TableRef improved in various places
  Upgrade to use core library version 0.80.4
  Added quoting of some variables
  simplified Xcode6-Beta4 release note
  @bmunkholm comment updates
  added link to all supported types in NSNumber property exception
  Updated README.md with license and pod tags
  clarified when CGFloat is a float vs a double
  Make message clearer
  updated NSNumber error message
  Using NSNumber properties (unsupported) now throws a more informative exception.
  - clarified CHANGELOG entry - removed “Note on float” - moved nsnumber_can_be_float() check to nsnumber_is_like_float()
  ...

+++ build.sh
@@ -15,7 +15,7 @@ set -o pipefail
 # You can override the version of the core library
 # Otherwise, use the default value
 if [ -z "$REALM_CORE_VERSION" ]; then
-    REALM_CORE_VERSION=0.80.3
+    REALM_CORE_VERSION=0.80.4
 fi
 
 PATH=/usr/local/bin:/usr/bin:/bin:/usr/libexec:$PATH
@@ -71,7 +71,7 @@ xcode5() {
 }
 
 xcode6() {
-    ln -s /Applications/Xcode6-Beta3.app/Contents/Developer/usr/bin build/bin || exit 1
+    ln -s /Applications/Xcode6-Beta4.app/Contents/Developer/usr/bin build/bin || exit 1
     PATH=./build/bin:$PATH xcodebuild -IDECustomDerivedDataLocation=build/DerivedData $@
 }
 
@@ -199,6 +199,10 @@ case "$COMMAND" in
     # Download Core Library
     ######################################
     "download-core")
+        if [ "$REALM_CORE_VERSION" = "current" ]; then
+            echo "Using version of core already in core/ directory"
+            exit 0
+        fi
         if ! [ -L core ]; then
             echo "core is not a symlink. Deleting..."
             rm -rf core
@@ -207,7 +211,6 @@ case "$COMMAND" in
             download_core
         else
             echo "The core library seems to be up to date."
-            echo "To force an update remove the folder 'core' and rerun."
         fi
         exit 0
         ;;
@@ -231,9 +234,9 @@ case "$COMMAND" in
         if [[ "$XCODE_VERSION" == "6" ]]; then
             # Build Universal Simulator/Device framework
             xcrealm "-scheme iOS -configuration Release -sdk iphonesimulator"
-            xcrealm "-scheme iOS -configuration Release"
+            xcrealm "-scheme iOS -configuration Release -sdk iphoneos"
             cd build/DerivedData/Realm-Xcode6/Build/Products || exit 1
-            mkdir -p Release || exit 1
+            mkdir -p Release-iphone || exit 1
             cp -R Release-iphoneos/Realm.framework Release-iphone || exit 1
             lipo -create -output Realm Release-iphoneos/Realm.framework/Realm Release-iphonesimulator/Realm.framework/Realm || exit 1
             mv Realm Release-iphone/Realm.framework || exit 1
@@ -260,9 +263,9 @@ case "$COMMAND" in
         if [[ "$XCODE_VERSION" == "6" ]]; then
             # Build Universal Simulator/Device framework
             xcrealm "-scheme iOS -configuration Debug -sdk iphonesimulator"
-            xcrealm "-scheme iOS -configuration Debug"
+            xcrealm "-scheme iOS -configuration Debug -sdk iphoneos"
             cd build/DerivedData/Realm-Xcode6/Build/Products || exit 1
-            mkdir -p Debug || exit 1
+            mkdir -p Debug-iphone || exit 1
             cp -R Debug-iphoneos/Realm.framework Debug-iphone || exit 1
             lipo -create -output Realm Debug-iphoneos/Realm.framework/Realm Debug-iphonesimulator/Realm.framework/Realm || exit 1
             mv Realm Debug-iphone/Realm.framework || exit 1
@@ -364,7 +367,7 @@ case "$COMMAND" in
         xc "-project objc/RealmSimpleExample/RealmSimpleExample.xcodeproj -scheme RealmSimpleExample -configuration Release clean build ${CODESIGN_PARAMS}"
         xc "-project objc/RealmTableViewExample/RealmTableViewExample.xcodeproj -scheme RealmTableViewExample -configuration Release clean build ${CODESIGN_PARAMS}"
         xc "-project objc/RealmMigrationExample/RealmMigrationExample.xcodeproj -scheme RealmMigrationExample -configuration Release clean build ${CODESIGN_PARAMS}"
-        xc "-project objc/RealmRestExample/RealmRestExample.xcodeproj -scheme RealmRestExample -configuration Release clean build ${CODESIGN_PARAMS}"
+        #xc "-project objc/RealmRestExample/RealmRestExample.xcodeproj -scheme RealmRestExample -configuration Release clean build ${CODESIGN_PARAMS}"
 
         # Not all examples can be built using Xcode 6
         if [[ "$XCODE_VERSION" != "6" ]]; then
@@ -382,7 +385,7 @@ case "$COMMAND" in
         xc "-project objc/RealmSimpleExample/RealmSimpleExample.xcodeproj -scheme RealmSimpleExample -configuration Debug clean build ${CODESIGN_PARAMS}"
         xc "-project objc/RealmTableViewExample/RealmTableViewExample.xcodeproj -scheme RealmTableViewExample -configuration Debug clean build ${CODESIGN_PARAMS}"
         xc "-project objc/RealmMigrationExample/RealmMigrationExample.xcodeproj -scheme RealmMigrationExample -configuration Debug clean build ${CODESIGN_PARAMS}"
-        xc "-project objc/RealmRestExample/RealmRestExample.xcodeproj -scheme RealmRestExample -configuration Debug clean build ${CODESIGN_PARAMS}"
+        #xc "-project objc/RealmRestExample/RealmRestExample.xcodeproj -scheme RealmRestExample -configuration Debug clean build ${CODESIGN_PARAMS}"
 
         # Not all examples can be built using Xcode 6
         if [[ "$XCODE_VERSION" != "6" ]]; then
